refactor(user) : ディレクトリ構造修正
