+++ spe-con/src/app/_user/_login/LoginModal.tsx
@@ -1,9 +1,10 @@
 'use client';
+
 import { useState } from 'react';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
-import { Button } from '@/app/common/components/layouts/ui/button';
+import { Button } from '@/components/atoms/Button';
 import {
   Form,
   FormControl,
@@ -11,89 +12,81 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from '@/app/common/components/layouts/ui/form';
-import { Input } from '@/app/common/components/layouts/ui/input';
-import { PasswordInput } from '../components/password-input';
+} from '@/components/atoms/Form';
+import { Input } from '@/components/atoms/Input';
+import { PasswordInput } from '../_components/PasswordInput';
 import { setCookie } from 'cookies-next';
-import { isLoginAtom } from '@/app/atoms/isLoginState';
+import { isLoginAtom } from '@/stores/isLoginState';
 import { useAtom } from 'jotai';
 import { useRouter } from 'next/navigation';
 import { ToastContainer, toast } from 'react-toastify';
 
 // フォームのバリデーションスキーマ
-const signupFormSchema = z
-  .object({
-    email: z.string().email({
-      message: '有効なメールアドレスを入力してください',
-    }),
-    password: z.string().min(8, {
-      message: 'パスワードは8文字以上で入力してください',
-    }),
-    password_confirmation: z
-      .string()
-      .min(8, { message: 'パスワード（確認）も8文字以上で入力してください' }),
-  })
-  .refine((data) => data.password === data.password_confirmation, {
-    message: 'パスワードが一致しません',
-    path: ['password_confirmation'],
-  });
+const loginFormSchema = z.object({
+  email: z.string().email({
+    message: '有効なメールアドレスを入力してください',
+  }),
+  password: z.string().min(1, {
+    message: 'パスワードを入力してください',
+  }),
+});
 
-type SignupFormValues = z.infer<typeof signupFormSchema>;
+type LoginFormValues = z.infer<typeof loginFormSchema>;
 
-interface SignupModalProps {
+interface LoginModalProps {
   onSuccess?: () => void;
-  onSwitchToLogin: () => void;
+  onSwitchToSignup: () => void;
+  onForgotPassword: () => void;
 }
 
-export function SignupModal({ onSuccess, onSwitchToLogin }: SignupModalProps) {
+
+export function LoginModal({ onSuccess, onSwitchToSignup, onForgotPassword }: LoginModalProps) {
   const [isLoading, setIsLoading] = useState(false);
   const [, setIsLogin] = useAtom(isLoginAtom);
   const router = useRouter();
 
   // フォームの初期化
-  const form = useForm<SignupFormValues>({
-    resolver: zodResolver(signupFormSchema),
+  const form = useForm<LoginFormValues>({
+    resolver: zodResolver(loginFormSchema),
     defaultValues: {
       email: '',
       password: '',
-      password_confirmation: '',
     },
   });
-
   const { setError } = form;
 
+ 
   // フォーム送信処理
-  async function onSubmit(value: SignupFormValues) {
+  async function onSubmit(values: LoginFormValues) {
     setIsLoading(true);
-console.log(`${process.env.NEXT_PUBLIC_API_URL}/auth`)
-console.log(`${process.env.NEXT_PUBLIC_NEXT_ENV}`)
+
     try {
-      const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth`, {
+      const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/sign_in`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
-        body: JSON.stringify(value),
+        body: JSON.stringify(values),
       });
 
       const data = await res.json();
-
+      
+      console.log(data)
       if (data.errors) {
         Object.entries(data.errors).forEach(([field, messages]) => {
-          setError(field as keyof SignupFormValues, {
+          setError(field as keyof LoginFormValues , {
             type: 'server',
             message: Array.isArray(messages) ? messages.join('、') : String(messages),
           });
         });
         return;
       }
+       setCookie('_access_token', res.headers.get('access-token'));
+            setCookie('_client', res.headers.get('client'));
+            setCookie('_uid', res.headers.get('uid'));
+            setIsLogin(true);
 
-      setCookie('_access_token', res.headers.get('access-token'));
-      setCookie('_client', res.headers.get('client'));
-      setCookie('_uid', res.headers.get('uid'));
-      setIsLogin(true);
-
-      toast.success('😇 登録完了', {
+      toast.success('😇 ログインしました', {
         position: 'top-center',
         autoClose: 1000,
         hideProgressBar: false,
@@ -106,11 +99,13 @@ console.log(`${process.env.NEXT_PUBLIC_NEXT_ENV}`)
       setTimeout(() => {
         router.push('/practice-menu');
       }, 2000);
-
+      // 成功したら3秒後にコールバック実行（デモ用）
       setTimeout(() => {
         if (onSuccess) onSuccess();
         form.reset();
-      }, 1000);
+      }, 3000);
+    } catch (error) {
+      console.error('ログインエラー:', error);
     } finally {
       setIsLoading(false);
     }
@@ -118,10 +113,10 @@ console.log(`${process.env.NEXT_PUBLIC_NEXT_ENV}`)
 
   return (
     <>
-      <ToastContainer />
+    <ToastContainer / >
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-3 sm:space-y-4 py-4">
-          <FormField
+        <FormField
             control={form.control}
             name="email"
             render={({ field }) => (
@@ -147,43 +142,40 @@ console.log(`${process.env.NEXT_PUBLIC_NEXT_ENV}`)
               <PasswordInput
                 name="password"
                 label="パスワード"
-                placeholder="8文字以上で入力"
+                placeholder="パスワードを入力"
                 field={field}
               />
             )}
           />
 
-          <FormField
-            control={form.control}
-            name="password_confirmation"
-            render={({ field }) => (
-              <PasswordInput
-                name="password_confirmation"
-                label="パスワード（確認）"
-                placeholder="パスワードを再入力"
-                field={field}
-              />
-            )}
-          />
+          <div className="flex justify-end">
+            <button
+              type="button"
+              onClick={onForgotPassword}
+              className="text-xs sm:text-sm text-[#2d7f98] hover:underline"
+            >
+              パスワードをお忘れですか？
+            </button>
+          </div>
 
           <Button
             type="submit"
             className="w-full bg-[#2d7f98] hover:bg-[#236a80] h-9 sm:h-10 text-sm sm:text-base mt-2"
             disabled={isLoading}
           >
-            {isLoading ? '登録中...' : 'アカウント登録'}
+            {isLoading ? 'ログイン中...' : 'ログイン'}
           </Button>
         </form>
       </Form>
 
       <div className="text-center mt-2 pb-6">
         <p className="text-gray-600 text-xs sm:text-sm">
-          すでにアカウントをお持ちですか？{' '}
+          アカウントをお持ちでないですか？{' '}
           <button
-            onClick={onSwitchToLogin}
+            onClick={onSwitchToSignup}
             className="text-[#2d7f98] font-semibold hover:underline"
           >
-            ログイン
+            新規登録
           </button>
         </p>
       </div>
