+++ spe-con/src/app/_user/_signup/SignupModal.tsx
@@ -1,10 +1,9 @@
 'use client';
-
 import { useState } from 'react';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
-import { Button } from '@/app/common/components/layouts/ui/button';
+import { Button } from '@/components/atoms/Button';
 import {
   Form,
   FormControl,
@@ -12,81 +11,89 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from '@/app/common/components/layouts/ui/form';
-import { Input } from '@/app/common/components/layouts/ui/input';
-import { PasswordInput } from '../components/password-input';
+} from '@/components/atoms/Form';
+import { Input } from '@/components/atoms/Input';
+import { PasswordInput } from '../_components/PasswordInput';
 import { setCookie } from 'cookies-next';
-import { isLoginAtom } from '@/app/atoms/isLoginState';
+import { isLoginAtom } from '@/stores/isLoginState';
 import { useAtom } from 'jotai';
 import { useRouter } from 'next/navigation';
 import { ToastContainer, toast } from 'react-toastify';
 
 // フォームのバリデーションスキーマ
-const loginFormSchema = z.object({
-  email: z.string().email({
-    message: '有効なメールアドレスを入力してください',
-  }),
-  password: z.string().min(1, {
-    message: 'パスワードを入力してください',
-  }),
-});
+const signupFormSchema = z
+  .object({
+    email: z.string().email({
+      message: '有効なメールアドレスを入力してください',
+    }),
+    password: z.string().min(8, {
+      message: 'パスワードは8文字以上で入力してください',
+    }),
+    password_confirmation: z
+      .string()
+      .min(8, { message: 'パスワード（確認）も8文字以上で入力してください' }),
+  })
+  .refine((data) => data.password === data.password_confirmation, {
+    message: 'パスワードが一致しません',
+    path: ['password_confirmation'],
+  });
 
-type LoginFormValues = z.infer<typeof loginFormSchema>;
+type SignupFormValues = z.infer<typeof signupFormSchema>;
 
-interface LoginModalProps {
+interface SignupModalProps {
   onSuccess?: () => void;
-  onSwitchToSignup: () => void;
-  onForgotPassword: () => void;
+  onSwitchToLogin: () => void;
 }
 
-
-export function LoginModal({ onSuccess, onSwitchToSignup, onForgotPassword }: LoginModalProps) {
+export function SignupModal({ onSuccess, onSwitchToLogin }: SignupModalProps) {
   const [isLoading, setIsLoading] = useState(false);
   const [, setIsLogin] = useAtom(isLoginAtom);
   const router = useRouter();
 
   // フォームの初期化
-  const form = useForm<LoginFormValues>({
-    resolver: zodResolver(loginFormSchema),
+  const form = useForm<SignupFormValues>({
+    resolver: zodResolver(signupFormSchema),
     defaultValues: {
       email: '',
       password: '',
+      password_confirmation: '',
     },
   });
+
   const { setError } = form;
 
- 
   // フォーム送信処理
-  async function onSubmit(values: LoginFormValues) {
+  async function onSubmit(value: SignupFormValues) {
     setIsLoading(true);
-
+console.log(`${process.env.NEXT_PUBLIC_API_URL}/auth`)
+console.log(`${process.env.NEXT_PUBLIC_NEXT_ENV}`)
     try {
-      const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/sign_in`, {
+      const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
-        body: JSON.stringify(values),
+        body: JSON.stringify(value),
       });
 
       const data = await res.json();
-      
-      console.log(data)
+
       if (data.errors) {
         Object.entries(data.errors).forEach(([field, messages]) => {
-          setError(field as keyof LoginFormValues , {
+          setError(field as keyof SignupFormValues, {
             type: 'server',
             message: Array.isArray(messages) ? messages.join('、') : String(messages),
           });
         });
         return;
       }
-       setCookie('_access_token', res.headers.get('access-token'));
-            setCookie('_client', res.headers.get('client'));
-            setCookie('_uid', res.headers.get('uid'));
-            setIsLogin(true);
 
-      toast.success('😇 ログインしました', {
+      setCookie('_access_token', res.headers.get('access-token'));
+      setCookie('_client', res.headers.get('client'));
+      setCookie('_uid', res.headers.get('uid'));
+      setIsLogin(true);
+
+      toast.success('😇 登録完了', {
         position: 'top-center',
         autoClose: 1000,
         hideProgressBar: false,
@@ -99,13 +106,11 @@ export function LoginModal({ onSuccess, onSwitchToSignup, onForgotPassword }: Lo
       setTimeout(() => {
         router.push('/practice-menu');
       }, 2000);
-      // 成功したら3秒後にコールバック実行（デモ用）
+
       setTimeout(() => {
         if (onSuccess) onSuccess();
         form.reset();
-      }, 3000);
-    } catch (error) {
-      console.error('ログインエラー:', error);
+      }, 1000);
     } finally {
       setIsLoading(false);
     }
@@ -113,10 +118,10 @@ export function LoginModal({ onSuccess, onSwitchToSignup, onForgotPassword }: Lo
 
   return (
     <>
-    <ToastContainer / >
+      <ToastContainer />
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-3 sm:space-y-4 py-4">
-        <FormField
+          <FormField
             control={form.control}
             name="email"
             render={({ field }) => (
@@ -142,40 +147,43 @@ export function LoginModal({ onSuccess, onSwitchToSignup, onForgotPassword }: Lo
               <PasswordInput
                 name="password"
                 label="パスワード"
-                placeholder="パスワードを入力"
+                placeholder="8文字以上で入力"
                 field={field}
               />
             )}
           />
 
-          <div className="flex justify-end">
-            <button
-              type="button"
-              onClick={onForgotPassword}
-              className="text-xs sm:text-sm text-[#2d7f98] hover:underline"
-            >
-              パスワードをお忘れですか？
-            </button>
-          </div>
+          <FormField
+            control={form.control}
+            name="password_confirmation"
+            render={({ field }) => (
+              <PasswordInput
+                name="password_confirmation"
+                label="パスワード（確認）"
+                placeholder="パスワードを再入力"
+                field={field}
+              />
+            )}
+          />
 
           <Button
             type="submit"
             className="w-full bg-[#2d7f98] hover:bg-[#236a80] h-9 sm:h-10 text-sm sm:text-base mt-2"
             disabled={isLoading}
           >
-            {isLoading ? 'ログイン中...' : 'ログイン'}
+            {isLoading ? '登録中...' : 'アカウント登録'}
           </Button>
         </form>
       </Form>
 
       <div className="text-center mt-2 pb-6">
         <p className="text-gray-600 text-xs sm:text-sm">
-          アカウントをお持ちでないですか？{' '}
+          すでにアカウントをお持ちですか？{' '}
           <button
-            onClick={onSwitchToSignup}
+            onClick={onSwitchToLogin}
             className="text-[#2d7f98] font-semibold hover:underline"
           >
-            新規登録
+            ログイン
           </button>
         </p>
       </div>
