+++ spe-con/src/app/_user/_types/modal.ts
