+++ spe-con/src/app/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { Home } from './home/Home';
+import { Home } from '../app/_home/Home';
 
 export default function Page() {
   return (
