+++ spe-con/src/types/index.ts
