+++ spe-con/src/utils/utils.ts
